requirements.txt
# LangChain 1.2
langchain>=1.2.0,<1.3.0
langchain-deepseek>=1.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.9

# 环境变量
python-dotenv>=1.0.0
